separate into platform layer
