+++ engine/engine.js
@@ -1,77 +1,31 @@
 import { bitmapTextToImageData } from "./bitmap.js";
-import { dispatch } from "../dispatch.js";
-import { textToTune } from '../textTuneConverters.js';
-import { global_state } from "../global_state.js";
-import { sizeGameCanvas } from "../dispatches/sizeGameCanvas.js";
-import * as render from "./render.js";
 
-
-let cur = null;
-
-export function init(canvas) {
-  // remove event listeners
-  let newCanvas = canvas.cloneNode(true);
-  canvas.parentNode.replaceChild(newCanvas, canvas);
-  canvas = newCanvas;
-
-  render.init(canvas);
-
-  canvas.setAttribute("tabindex", "1");
-
-
-  function gameloop() {
-    render.render(drawTiles());
-
-    animationId = window.requestAnimationFrame(gameloop);
+// Tagged tempalate literal factory go brrr
+export function _makeTag(cb) {
+  return (strings, ...interps) => {
+    if (typeof strings === "string") {
+      throw new Error("Tagged template literal must be used like name`text`, instead of name(`text`)");
+    }
+    const string = strings.reduce((p, c, i) => p + c + (interps[i] ?? ''), '');
+    return cb(string);
   }
+}
 
-  function end() {
-    window.cancelAnimationFrame(animationId);
-  }
-
-  if (cur) cur();
-  cur = end;
-
-
-   let animationId = window.requestAnimationFrame(gameloop);
-
-  function setScreenSize(w, h) {
-    canvas.width = w;
-    canvas.height = h;
-
-    window.idealDimensions = [dimensions.width, dimensions.height];
-    sizeGameCanvas();
-
-    render.resize(canvas);
-  }
-
+export function init({ palette, setBitmaps, setScreenSize }) {
   // tile gamelab
-  let legend = [];
-  let dimensions = {
-    width: 0,
-    height: 0,
-    maxTileDim: 0,
-  }
-  let sprites = [];
-  let tileInputs = {
-    up: [],
-    down: [],
-    left: [],
-    right: [],
-    i: [],
-    j: [],
-    k: [],
-    l: [],
-  };
-  let afterInputs = [];
-  let solids = [];
-  let pushable = {};
-  let maxTileDim = 0;
 
-  let background = "";
-  const tempCanvas = document.createElement("canvas");
-  // tempCanvas.width = 16;
-  // tempCanvas.height = 16;
+  const state = {
+    legend: [],
+    dimensions: {
+      width: 0,
+      height: 0,
+      maxTileDim: 0,
+    },
+    sprites: [],
+    solids: [],
+    pushable: {},
+    maxTileDim: 0,
+  };
 
   class Sprite {
     constructor(type, x, y) {
@@ -83,7 +37,7 @@ export function init(canvas) {
     }
 
     set type(k) {
-      const legendDict = Object.fromEntries(legend);
+      const legendDict = Object.fromEntries(state.legend);
       if (!(k in legendDict)) throw new Error(`"${k}" not in legend.`);
 
       this.remove();
@@ -115,15 +69,14 @@ export function init(canvas) {
     }
 
     remove() {
-      sprites = sprites.filter(s => s !== this);
+      state.sprites = state.sprites.filter(s => s !== this);
       return this;
     }
   }
 
-
   const _canMoveToPush = (sprite, dx, dy) => {
     const { x, y, type } = sprite;
-    const { width, height } = dimensions;
+    const { width, height } = state.dimensions;
     const i = (x+dx)+(y+dy)*width;
 
     const inBounds = (x+dx < width && x+dx >= 0 && y+dy < height && y+dy >= 0);
@@ -131,7 +84,7 @@ export function init(canvas) {
 
     const grid = getGrid();
 
-    const notSolid = !solids.includes(type);
+    const notSolid = !state.solids.includes(type);
     const noMovement = dx === 0 && dy === 0;
     const movingToEmpty = i < grid.length && grid[i].length === 0;
 
@@ -144,7 +97,7 @@ export function init(canvas) {
     let canMove = true;
 
     grid[i].forEach(sprite => {
-      const isSolid = solids.includes(sprite.type);
+      const isSolid = state.solids.includes(sprite.type);
       const isPushable = (type in pushable) && pushable[type].includes(sprite.type);
 
       if (isSolid && !isPushable)
@@ -163,37 +116,11 @@ export function init(canvas) {
     return canMove;
   }
 
-  canvas.addEventListener("keydown", (e) => {
-    const key = e.key;
-
-    const VALID_INPUTS = ["w", "a", "s", "d", "i", "j", "k", "l"];
-
-    if (!VALID_INPUTS.includes(key)) return;
-
-    if (key === "w") tileInputs["up"].forEach(fn => fn());
-    if (key === "a") tileInputs["left"].forEach(fn => fn());
-    if (key === "s") tileInputs["down"].forEach(fn => fn());
-    if (key === "d") tileInputs["right"].forEach(fn => fn());
-    if (key === "i") tileInputs["i"].forEach(fn => fn());
-    if (key === "j") tileInputs["j"].forEach(fn => fn());
-    if (key === "k") tileInputs["k"].forEach(fn => fn());
-    if (key === "l") tileInputs["l"].forEach(fn => fn());
-
-    afterInputs.forEach(f => f());
-
-    sprites.forEach(s => {
-      s.dx = 0;
-      s.dy = 0;
-    })
-
-    e.preventDefault();
-  });
-
   const getGrid = () => {
-    const { width, height } = dimensions;
+    const { width, height } = state.dimensions;
 
     const grid = new Array(width*height).fill(0).map(x => []);
-    sprites.forEach(s => {
+    state.sprites.forEach(s => {
       const i = s.x+s.y*width;
       grid[i].push(s);
     })
@@ -202,13 +129,14 @@ export function init(canvas) {
   }
 
   const _checkBounds = (x, y) => {
-    const { width, height } = dimensions;
+    const { width, height } = state.dimensions;
 
     if (x > width || x < 0 || y < 0 || y > height) throw new Error(`Sprite out of bounds.`);
   }
 
   const _checkLegend = type => {
-    if (!(type in Object.fromEntries(legend))) throw new Error(`Unknown sprite type: ${type}`);
+    if (!(type in Object.fromEntries(state.legend)))
+      throw new Error(`Unknown sprite type: ${type}`);
   }
 
   const addSprite = (x, y, type) => {
@@ -218,55 +146,14 @@ export function init(canvas) {
     _checkLegend(type);
 
     const s = new Sprite(type, x, y);
-    sprites.push(s);
+    state.sprites.push(s);
   }
-
-  let cachedTileImages = {};
 
   function setLegend(...bitmaps) {
-    legend = bitmaps;
-
-    // check that legend keys are single characters
-
-    cachedTileImages = {};
-    render.setBitmaps(bitmaps);
-    dispatch("SET_BITMAPS", { bitmaps });
+    state.legend = bitmaps;
+    setBitmaps(bitmaps);
   }
-
   
-  function _getTileImage(type) {
-    const legendDict = Object.fromEntries(legend);
-    if (!(type in legendDict)) throw new Error(`Type not in legend: ${type}`);
-
-    const val = legendDict[type];
-    const isCombo = val.type !== undefined;
-
-    if (isCombo) throw new Error(`Can not draw combination sprites.`);
-
-    if (!(type in cachedTileImages)) {
-      const c = document.createElement("canvas");
-      c.width = val.imageData.width;
-      c.height = val.imageData.height;
-
-      const ctx = c.getContext("2d");
-
-      ctx.webkitImageSmoothingEnabled = false;
-      ctx.mozImageSmoothingEnabled = false;
-      ctx.imageSmoothingEnabled = false;
-
-      ctx.putImageData(
-        val.imageData, 
-        0,
-        0,
-      );
-
-      cachedTileImages[type] = c;
-    }
-
-    
-    return cachedTileImages[type];
-  }
-
   const _allEqual = arr => arr.every(val => val === arr[0]);
 
   function setMap(string) { 
@@ -276,10 +163,10 @@ export function init(canvas) {
     if (!isRect) throw new Error("Level must be rect.");
     const w = rows[0].length;
     const h = rows.length;
-    dimensions.width = w;
-    dimensions.height = h;
+    state.dimensions.width = w;
+    state.dimensions.height = h;
 
-    sprites = [];
+    state.sprites = [];
 
     setScreenSize(w*16, h*16);
 
@@ -298,11 +185,11 @@ export function init(canvas) {
   }
 
   function clearTile(x, y) {
-    sprites = sprites.filter(s => s.x !== x || s.y !== y);
+    state.sprites = state.sprites.filter(s => s.x !== x || s.y !== y);
   }
 
   function getTile(x, y) { 
-    return getGrid()[dimensions.width*y+x] || [];
+    return getGrid()[state.dimensions.width*y+x] || [];
   }
 
   function hasDuplicates(array) {
@@ -310,11 +197,12 @@ export function init(canvas) {
   }
 
   function tilesWith(...matchingTypes) {
+    const { width, height } = state.dimensions;
     const tiles = [];
     const grid = getGrid();
-    for (let x = 0; x < dimensions.width; x++) {
-      for (let y = 0; y < dimensions.height; y++) {
-        const tile = grid[dimensions.width*y+x] || [];
+    for (let x = 0; x < width; x++) {
+      for (let y = 0; y < height; y++) {
+        const tile = state.grid[width*y+x] || [];
         const matchIndices = matchingTypes.map(type => {
           return tile.map(s => s.type).indexOf(type);
         })
@@ -328,61 +216,14 @@ export function init(canvas) {
   }
 
   function setSolids(arr) {
-    solids = arr;
+    state.solids = arr;
   }
 
   function setPushables(map) {
-    pushable = map;
+    state.pushable = map;
   }
 
-  function onInput(type, fn) {
-    if (!(type in tileInputs)) console.error("Unknown input type:", type)
-    tileInputs[type].push(fn);
-  }
-
-  function drawTiles() {
-    const grid = getGrid();
-    const { width, height, maxTileDim } = dimensions;
-    if (width == 0 || height == 0) return new ImageData(1, 1);
-    const img = new ImageData(width, height);
-
-    for (let i = 0; i < grid.length; i++) {
-      const x = i%width; 
-      const y = Math.floor(i/width); 
-
-      const sprites = grid[i];
-      const zOrder = legend.map(x => x[0]);
-      sprites.sort((a, b) => zOrder.indexOf(a.type) - zOrder.indexOf(b.type));
-
-      for (let i = 0; i < 4; i++) {
-        if (!sprites[i]) continue;
-        const { type: t } = sprites[i];
-        img.data[(y*dimensions.width + x)*4 + i] = 1+legend.findIndex(f => f[0] == t);
-      }
-    }
-
-    return img;
-  }
-
-  function afterInput(fn) {
-    afterInputs.push(fn);
-  }
-
-  // how to add timed things, like bird flying and ball kicks
-  
-  // Tagged tempalate literal factory go brrr
-  function _makeTag(cb) {
-    return (strings, ...interps) => {
-      if (typeof strings === "string") {
-        throw new Error("Tagged template literal must be used like name`text`, instead of name(`text`)");
-      }
-      const string = strings.reduce((p, c, i) => p + c + (interps[i] ?? ''), '');
-      return cb(string);
-    }
-  }
-
-  return {
-    setScreenSize,
+  const api = {
     setLegend, 
     setMap, 
     addSprite,
@@ -390,20 +231,22 @@ export function init(canvas) {
     getTile,
     tilesWith,
     clearTile, 
-    onInput, 
     setSolids, 
     setPushables, 
-    afterInput, 
     map: _makeTag(text => text), // No-op for now, here for editor support
-    tune: _makeTag(text => textToTune(text)),
-    bitmap: _makeTag(text => ({text, imageData: bitmapTextToImageData(text, global_state.palette)})),
-    getFirst: (type) => sprites.find(t => t.type === type), // **
-    getAll: (type) => type ? sprites.filter(t => t.type === type) : sprites, // **
-    width: () => dimensions.width,
-    height: () => dimensions.height,
+    bitmap: _makeTag(text => ({
+      text,
+      imageData: bitmapTextToImageData(text, palette)
+    })),
+    getFirst: (type) => state.sprites.find(t => t.type === type), // **
+    getAll: (type) => type ? state.sprites.filter(t => t.type === type) : state.sprites, // **
+    width: () => state.dimensions.width,
+    height: () => state.dimensions.height,
     setBackground: (type) => { 
       _checkLegend(type);
       background = type;
     }
-  }
+  };
+
+  return { api, state };
 }
